changed deliveryPay and totalPay to float

+++ user_lib/src/main/java/com/crumbs/lib/entity/Driver.java
@@ -43,7 +43,7 @@ public class Driver {
         @ToString.Exclude
         private UserDetails userDetails;
 
-        private Long totalPay;
+        private Float totalPay;
 
 }
